+++ RELEASE-NOTES.md
@@ -16,6 +16,7 @@
 * #337 Compacting Parquet columns on load
 * #338 Compacting JDBC columns on load
 * #341 JsonLoader: align API naming with other loaders
+* #343 Compacting Excel columns on load
 
 ## Release 1.0.0-M23